+++ Changelog.md
@@ -1,5 +1,10 @@
 # Unreleased
 
+- Rewrite crate
+    - Add support for dependencies not named "bootloader"
+    - Use `resolve` table for detection to also support renamed dependencies
+    - Remove most dependencies -> improves compile time
+
 # 0.0.3 – 2020-08-24
 
 - Remove detection of bootloader features